pyproject.toml
[project]
name = "riddler"
version = "0.1.0"
description = ""
authors = [
    {name = "Rohan Sarvaria",email = "rsarvaria@uwaterloo.ca"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv (>=1.1.0,<2.0.0)",
    "discord (>=2.3.2,<3.0.0)",
    "toml (>=0.10.2,<0.11.0)",
    "dotmap (>=1.3.30,<2.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]

[tool.poetry]
packages = [{include = "riddler", from = "src"}]

[tool.poetry.scripts]
riddler = "src.__main__:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
